assignment/requirements.txt
pip==22.0.3
setuptools==60.6.0
wheel==0.37.1
django==3.2.12
djangorestframework
markdown
django-filter
requests
django-test-plus
black